Image Generation (initial)
Initial support for Image Generation of the classifier's decision function for an observation.

This feature is enabled by the IMAGE_GENERATION setting and the fields to be graphed are set with the IMAGE_GENERATION_FIELDS setting.

It will generate a graph with the pairwise conditional decision function with the fields specified in IMAGE_GENERATION_FIELDS.

The decision function is conditional to the other fields in the observation - i.e. if {rbc: 3, wbc:2, hgb:120}, the graph on the decision function on (rbc, wbc) is computed with hgb:120.

*CAVEATS*

- CHTUID_USE_IN_CLASSIFICATION is not supported (must be set to False and perform inference on the classifier to reflect the change if needed)
- All the fields / columns to be graphed must be present on the form, otherwise the graph won't be shown
- LEARNING_FIELDS must be equal to IMAGE_GENERATION_FIELDS, this can be set on the admin of the technique used for classification (performing inference may needed)
- LEARNING_FIELDS_CATEGORICAL is not supported. This can be set on the admin of the technique using the keyword "__none__" (re-training may needed)
- No testing coverage (experimental)
- Not enabled by default